requirements.txt
numpy == 1.26.4
opencv-python==4.10.0.82
keras==3.7.0
tensorflow==2.16.1
pillow==11.0.0
gradio==5.9.1